moved configuration files in their own directory

+++ install_lalinference_cbc.sh
@@ -22,46 +22,47 @@ if [ ! -d ~/pe/src/jhbuild ]; then
 fi
 
 # Set up jhbuild configuration files
-mkdir -p ~/.config && cd ~/.config
-ln -s ~/pe/modulesets/jhbuildrc
+mkdir -p ~/pe/.config && cd ~/pe/.config
+if [ ! -e ~/pe/.config/jhbuildrc ]; then
+  ln -s ~/pe/modulesets/jhbuildrc
+fi
 
-echo "prefix = '$HOME/pe/local/'" > ~/pe/modulesets/pe.jhbuildrc
-echo "modulesets_dir = '$HOME/pe/modulesets'" >> ~/pe/modulesets/pe.jhbuildrc
-echo "checkoutroot = '$HOME/pe/src'" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lal']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalframe']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalmetaio']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalxml']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalburst']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalpulsar']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalstochastic']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalinspiral']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalsimulation']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalinference']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalapps']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lalsuite-extra']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['glue']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['pylal']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['ligo']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['gracedb']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "branches['lvalert']=(None,'lalinference_o2')" >> ~/pe/modulesets/pe.jhbuildrc
-echo "" >> ~/pe/modulesets/pe.jhbuildrc
-echo "intel_executables=['icc','icpc','ifort','mpiicc','mpiicpc','mpiifort','xiar']" >> ~/pe/modulesets/pe.jhbuildrc
-echo "" >> ~/pe/modulesets/pe.jhbuildrc
-echo "from distutils.spawn import find_executable" >> ~/pe/modulesets/pe.jhbuildrc
-echo "def is_in_path(name):" >> ~/pe/modulesets/pe.jhbuildrc
-echo '    """Check whether `name` is on PATH."""' >> ~/pe/modulesets/pe.jhbuildrc
-echo "    return find_executable(name) is not None" >> ~/pe/modulesets/pe.jhbuildrc
-echo "" >> ~/pe/modulesets/pe.jhbuildrc
-echo "if all([is_in_path(name) for name in intel_executables]):" >> ~/pe/modulesets/pe.jhbuildrc
-echo "   icc = True" >> ~/pe/modulesets/pe.jhbuildrc
-echo "" >> ~/pe/modulesets/pe.jhbuildrc
-echo "del name" >> ~/pe/modulesets/pe.jhbuildrc
-echo "del intel_executables" >> ~/pe/modulesets/pe.jhbuildrc
-echo "" >> ~/pe/modulesets/pe.jhbuildrc
-echo "os.environ['LAL_DATA_PATH'] = '$HOME/ROM_data/'" >> ~/pe/modulesets/pe.jhbuildrc
-echo "" >> ~/pe/modulesets/pe.jhbuildrc
-ln -s ~/pe/modulesets/pe.jhbuildrc
+echo "prefix = '$HOME/pe/local/'" > ~/pe/.config/pe.jhbuildrc
+echo "modulesets_dir = '$HOME/pe/modulesets'" >> ~/pe/.config/pe.jhbuildrc
+echo "checkoutroot = '$HOME/pe/src'" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lal'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalframe'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalmetaio'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalxml'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalburst'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalpulsar'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalstochastic'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalinspiral'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalsimulation'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalinference'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalapps'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lalsuite-extra'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['glue'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['pylal'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['ligo'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['gracedb'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "branches['lvalert'] = (None,'lalinference_o2')" >> ~/pe/.config/pe.jhbuildrc
+echo "" >> ~/pe/.config/pe.jhbuildrc
+echo "intel_executables = ['icc','icpc','ifort','mpiicc','mpiicpc','mpiifort','xiar']" >> ~/pe/.config/pe.jhbuildrc
+echo "" >> ~/pe/.config/pe.jhbuildrc
+echo "from distutils.spawn import find_executable" >> ~/pe/.config/pe.jhbuildrc
+echo "def is_in_path(name):" >> ~/pe/.config/pe.jhbuildrc
+echo '    """Check whether `name` is on PATH."""' >> ~/pe/.config/pe.jhbuildrc
+echo "    return find_executable(name) is not None" >> ~/pe/.config/pe.jhbuildrc
+echo "" >> ~/pe/.config/pe.jhbuildrc
+echo "if all([is_in_path(name) for name in intel_executables]):" >> ~/pe/.config/pe.jhbuildrc
+echo "   icc = True" >> ~/pe/.config/pe.jhbuildrc
+echo "" >> ~/pe/.config/pe.jhbuildrc
+echo "del name" >> ~/pe/.config/pe.jhbuildrc
+echo "del intel_executables" >> ~/pe/.config/pe.jhbuildrc
+echo "" >> ~/pe/.config/pe.jhbuildrc
+echo "os.environ['LAL_DATA_PATH'] = '$HOME/ROM_data/'" >> ~/pe/.config/pe.jhbuildrc
+echo "" >> ~/pe/.config/pe.jhbuildrc
 
 # Install lalsuite from anonymous repository
 ~/pe/.local/bin/jhbuild build lalsuite
@@ -72,5 +73,5 @@ ln -s ~/pe/modulesets/pe.jhbuildrc
 
 # Create initisalisation script
 echo "#!/bin/bash" > ~/pe/lalinference_o2.sh
-echo "$HOME/pe/.local/bin/jhbuild -f $HOME/.config/jhbuildrc run $SHELL --noprofile --norc" >> ~/pe/lalinference_o2.sh
+echo "$HOME/pe/.local/bin/jhbuild -f $HOME/pe/.config/jhbuildrc run $SHELL --noprofile --norc" >> ~/pe/lalinference_o2.sh
 chmod a+x ~/pe/lalinference_o2.sh
